+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed magazines dropping when they no longer fit the slot the new magazine came from
 * Magazines are now only retained once the weapon's mastering is maxed out
 * Added a setting to turn the mastering requirement off
+* Added a setting to get notified when a magazine is dropped
 * Added a setting to disable magazine retention
 
 ## v1.2.0